requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
reportlab==4.2.2
python-docx==1.1.2

openai==1.40.6
requests==2.32.3

numpy==2.0.1
hnswlib==0.8.0
sentence-transformers==3.0.1
torch>=2.1.0